Start a base class for thinkers that use ML
