+++ examol/steer/single.py
@@ -1,11 +1,4 @@
 """Single-objective and single-fidelity implementation of active learning. As easy as we get"""
-import gzip
-import json
-import os
-import pickle as pkl
-import shutil
-from concurrent.futures import ProcessPoolExecutor
-from functools import partial
 from pathlib import Path
 from queue import Queue
 from threading import Event
@@ -15,21 +8,19 @@ from typing import Sequence
 import numpy as np
 from colmena.queue import ColmenaQueues
 from colmena.thinker import event_responder, ResourceCounter, agent
-from more_itertools import interleave_longest, batched
+from more_itertools import interleave_longest
 from proxystore.proxy import Proxy, extract
-from proxystore.store import Store, get_store
 from proxystore.store.base import ConnectorKeyT
 from proxystore.store.utils import get_key
 
-from .base import MoleculeThinker
-from ..score.base import Scorer
+from .base import ScorerThinker
 from examol.solution import SingleFidelityActiveLearning
 from ..store.db.base import MoleculeStore
 from ..store.models import MoleculeRecord
 from ..store.recipes import PropertyRecipe
 
 
-class SingleStepThinker(MoleculeThinker):
+class SingleStepThinker(ScorerThinker):
     """A thinker which submits all computations needed to evaluate a molecule whenever it is selected
 
     Args:
@@ -41,11 +32,6 @@ class SingleStepThinker(MoleculeThinker):
         search_space: Search space of molecules. Provided as a list of paths to ".smi" files
         num_workers: Number of simulation tasks to run in parallel
         inference_chunk_size: Number of molecules to run inference on per task
-
-    Attributes:
-        search_space_smiles: SMILES associated with each molecule in the search space, broken into chunks
-        search_space_inputs: Inputs to the ML models for each molecule in the search space, broken into chucks
-        database: Map between molecule InChI key and currently-known information about it
     """
 
     def __init__(self,
@@ -57,11 +43,10 @@ class SingleStepThinker(MoleculeThinker):
                  search_space: list[Path | str],
                  num_workers: int = 2,
                  inference_chunk_size: int = 10000):
-        super().__init__(queues, ResourceCounter(num_workers), run_dir, recipes, solution, search_space, database)
+        super().__init__(queues, ResourceCounter(num_workers), run_dir, recipes, solution.scorer, solution, search_space, database, inference_chunk_size)
 
         # Store the selection equipment
         self.solution = solution
-        self.scorer = self.solution.scorer
         self.models = self.solution.models.copy()
         self.selector = self.solution.selector
         self.starter = self.solution.starter
@@ -69,12 +54,6 @@ class SingleStepThinker(MoleculeThinker):
             raise ValueError('You must provide the same number of models for each class')
         if len(self.models) != len(recipes):  # pragma: no-coverage
             raise ValueError('You must provide as many model ensembles as recipes')
-
-        # Partition the search space into smaller chunks
-        self.search_space_dir = self.run_dir / 'search-space'
-        self.search_space_smiles: list[list[str]]
-        self.search_space_inputs: list[list[object]]
-        self.search_space_smiles, self.search_space_inputs = zip(*self._cache_search_space(inference_chunk_size, self.search_space))
 
         # Model tracking information
         self._model_proxy_keys: list[list[ConnectorKeyT | None]] = [[None] * len(m) for m in self.models]  # Proxy for the current model
@@ -88,82 +67,6 @@ class SingleStepThinker(MoleculeThinker):
     def num_models(self) -> int:
         """Number of models being trained by this class"""
         return sum(map(len, self.models))
-
-    @property
-    def inference_store(self) -> Store | None:
-        """Proxystore used for inference tasks"""
-        if (store_name := self.queues.proxystore_name.get('inference')) is not None:
-            return get_store(store_name)
-
-    def _cache_search_space(self, inference_chunk_size: int, search_space: list[str | Path]):
-        """Cache the search space into a directory within the run"""
-
-        # Check if we must rebuild the cache
-        rebuild = True
-        config_path = self.search_space_dir / 'settings.json'
-        my_config = {
-            'inference_chunk_size': inference_chunk_size,
-            'scorer': str(self.scorer),
-            'paths': [str(Path(p).resolve()) for p in search_space]
-        }
-        if config_path.exists():
-            config = json.loads(config_path.read_text())
-            rebuild = config != my_config
-            if rebuild:
-                self.logger.info('Settings have changed. Rebuilding the cache')
-                shutil.rmtree(self.search_space_dir)
-        elif self.search_space_dir.exists():
-            shutil.rmtree(self.search_space_dir)
-        self.search_space_dir.mkdir(exist_ok=True, parents=True)
-
-        # Get the paths to inputs and keys, either by rebuilding or reading from disk
-        search_space_keys = {}
-        if rebuild:
-            # Build search space and save to disk
-
-            # Process the inputs and store them to disk
-            search_size = 0
-            input_func = partial(_generate_inputs, scorer=self.scorer)
-            with ProcessPoolExecutor(min(4, os.cpu_count())) as pool:
-                mol_iter = pool.map(input_func, self.iterate_over_search_space(), chunksize=1000)
-                mol_iter_no_failures = filter(lambda x: x is not None, mol_iter)
-                for chunk_id, chunk in enumerate(batched(mol_iter_no_failures, inference_chunk_size)):
-                    keys, objects = zip(*chunk)
-                    search_size += len(keys)
-                    chunk_path = self.search_space_dir / f'chunk-{chunk_id}.pkl.gz'
-                    with gzip.open(chunk_path, 'wb') as fp:
-                        pkl.dump(objects, fp)
-
-                    search_space_keys[chunk_path.name] = keys
-            self.logger.info(f'Saved {search_size} search entries into {len(search_space_keys)} batches')
-
-            # Save the keys and the configuration
-            with open(self.search_space_dir / 'keys.json', 'w') as fp:
-                json.dump(search_space_keys, fp)
-            with config_path.open('w') as fp:
-                json.dump(my_config, fp)
-        else:
-            # Load in keys
-            self.logger.info(f'Loading search space from {self.search_space_dir}')
-            with open(self.search_space_dir / 'keys.json') as fp:
-                search_space_keys = json.load(fp)
-
-        # Load in the molecules, storing them as proxies in the "inference" store if there is a store defined
-        self.logger.info(f'Loading in molecules from {len(search_space_keys)} files')
-        output = []
-
-        proxy_store = self.inference_store
-        if proxy_store is not None:
-            self.logger.info(f'Will store inference objects to {proxy_store}')
-
-        for name, keys in search_space_keys.items():
-            with gzip.open(self.search_space_dir / name, 'rb') as fp:  # Load from disk
-                objects = pkl.load(fp)
-
-            if proxy_store is not None:  # If the store exists, make a proxy
-                objects = proxy_store.proxy(objects)
-            output.append((keys, objects))
-        return output
 
     def _simulations_complete(self):
         self.start_training.set()
@@ -345,23 +248,3 @@ class SingleStepThinker(MoleculeThinker):
             List of molecules for which that property is defined
         """
         return [x for x in self.database.iterate_over_records() if recipe.lookup(x) is not None]
-
-
-def _generate_inputs(record: MoleculeRecord, scorer: Scorer) -> tuple[str, object] | None:
-    """Parse a molecule then generate a form ready for inference
-
-    Args:
-        record: Molecule record to be parsed
-        scorer: Tool used for inference
-    Returns:
-        - Key for the molecule record
-        - Inference-ready format
-        Or None if the transformation fails
-    """
-
-    try:
-        # Compute the features
-        readied = scorer.transform_inputs([record])[0]
-    except (ValueError, RuntimeError):
-        return None
-    return record.identifier.smiles, readied
